src/requirements.txt
python-telegram-bot==21.5
openai==1.55.3
python-dotenv==1.0.1
python-telegram-bot[job-queue]
sentry-sdk>=2.8.0
httpx==0.27.2